Separate classes into new files
